Refactor ros2_ws image and scripts
* Add sshd_config file to Dockerfile

* Streamline build.sh script and support MacOS

* Add server.sh script with arg parsing to spinup
a built image as a background SSH server.

* Add connect.sh script as a shortcut for 'exec'
attaching an interactive shell to a running
container, with platform specific environment
variables and volumes for display configuration.

* Add apt-key file downloaded from
http://repo.ros2.org/repos.key
to ros2_ws/config directory to add with apt-key in
the Dockerfile. Note: curl is not used because
it is not yet installed in the base ros image,
and cannot be installed with apt without the key!

* Consolidate contents of ros2_ws config directory into
Dockerfile directly (99_aptget and update_bashrc),
both to reduce clutter and to allow passing the USER
variable instead of hardcoding ros2

* Reduce the amount of setup sourcing done in the various
bashrc and entrypoint scripts in the Dockerfile, simply
sourcing the core ros and ros2_ws setup scripts at the top
of bashrc